+++ packages/remix-eslint-config/rules/import.js
@@ -5,5 +5,6 @@ const ERROR = 2;
 module.exports = {
   "import/first": ERROR,
   "import/no-amd": ERROR,
+  "import/no-duplicates": ERROR,
   "import/no-webpack-loader-syntax": ERROR,
 };