+++ packages/remix-react/index.tsx
@@ -24,6 +24,7 @@ export {
   useFormAction,
   useHref,
   useLocation,
+  useMatch,
   useNavigate,
   useNavigation,
   useNavigationType,