+++ packages/remix-serve/CHANGELOG.md
@@ -1,5 +1,62 @@
 # `@remix-run/serve`
 
+## 1.18.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.18.1`
+  - `@remix-run/express@1.18.1`
+
+## 1.18.0
+
+### Minor Changes
+
+- stabilize v2 dev server ([#6615](https://github.com/remix-run/remix/pull/6615))
+
+### Patch Changes
+
+- fix(types): better tuple serialization types ([#6616](https://github.com/remix-run/remix/pull/6616))
+- Updated dependencies:
+  - `@remix-run/node@1.18.0`
+  - `@remix-run/express@1.18.0`
+
+## 1.17.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/express@1.17.1`
+  - `@remix-run/node@1.17.1`
+
+## 1.17.0
+
+### Patch Changes
+
+- Add `HeadersArgs` type to be consistent with loaders/actions/meta and allows for using a `function` declaration in addition to an arrow function expression ([#6247](https://github.com/remix-run/remix/pull/6247))
+
+  ```tsx
+  import type { HeadersArgs } from "@remix-run/node"; // or cloudflare/deno
+
+  export function headers({ loaderHeaders }: HeadersArgs) {
+    return {
+      "x-my-custom-thing": loaderHeaders.get("x-my-custom-thing") || "fallback",
+    };
+  }
+  ```
+
+- Updated dependencies:
+  - `@remix-run/node@1.17.0`
+  - `@remix-run/express@1.17.0`
+
+## 1.16.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.16.1`
+  - `@remix-run/express@1.16.1`
+
 ## 1.16.0
 
 ### Patch Changes